little cleanup

+++ src/backtrace.c
@@ -9,7 +9,7 @@ int64_t getAdressE8(int64_t operand_begin, int64_t base){
 }
 
 int checkE8(int64_t after_call){
-	int64_t temp = (int64_t)*((int64_t *)(after_call - 5));
+	int64_t temp = *((int64_t *)(after_call - 5));
 	return (temp & (0xFF)) == 0xE8;
 }
 
@@ -18,22 +18,22 @@ void backtrace(int64_t *buf, int size, int iret){
 	__asm__ volatile("mov %%rbp, %0" : "=r"(rbp));
 	printf("backtrace begin=============================\n\n");
 	for(int i = 0; i < size; i++, iret--, rbp = *((int64_t *)rbp)){
-			int64_t after_call = (int64_t)*((int64_t *)(rbp + 8));
-			if(!iret){
-				printf("IRET TO [%p]\n\n", after_call);
-				continue;
-			}
+		int64_t after_call = *((int64_t *)(rbp + 8));
+		if(!iret){
+			printf("IRET TO [%p]\n\n", after_call);
+			continue;
+		}
 
-			printf("CALLER POINT ADRESS (after call) [%p]\n", after_call);
+		printf("CALLER POINT ADRESS (after call) [%p]\n", after_call);
 
-			if(!checkE8(after_call)){
-				printf("UNSUPPORTED CALL FORMAT\n\n");
-				break;
-			}
+		if(!checkE8(after_call)){
+			printf("UNSUPPORTED CALL FORMAT\n\n");
+			break;
+		}
 
-			int64_t temp = (int64_t)*((int64_t *)(after_call - 4));
-			buf[i] = getAdressE8(temp, after_call);
-			printf("CURRENT ADRESS CALL [%p]\n\n", buf[i]);
+		int64_t temp = *((int64_t *)(after_call - 4));
+		buf[i] = getAdressE8(temp, after_call);
+		printf("CURRENT ADRESS CALL [%p]\n\n", buf[i]);
 	}
 	printf("backtrace end=============================\n");
 }
